feat: add role column to users table and user_id foreign key to academic_years table

+++ database/migrations/2026_07_20_050246_add_user_id.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('academic_years', function (Blueprint $table) {
-            $table->foreignId('user_id')->nullable()->constraind('users')->cascadeOnDelete();
-        });
+        if (! Schema::hasColumn('academic_years', 'user_id')) {
+            Schema::table('academic_years', function (Blueprint $table) {
+                $table->foreignId('user_id')->nullable()->constrained('users')->cascadeOnDelete();
+            });
+        }
     }
 
     /**
@@ -21,8 +23,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('academic_years', function (Blueprint $table) {
-            $table->dropColumn('user_id');
-        });
+        if (Schema::hasColumn('academic_years', 'user_id')) {
+            Schema::table('academic_years', function (Blueprint $table) {
+                $table->dropForeign(['user_id']);
+                $table->dropColumn('user_id');
+            });
+        }
     }
 };
